Conflicts:
	BussinessLogic/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	Data/bin/x86/Debug/ProntoMVC.Data.dll
	Data/bin/x86/Debug/ProntoMVC.Data.pdb
	Data/obj/x86/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	Data/obj/x86/Debug/ProntoMVC.Data.csproj.FileListAbsolute.txt
	Data/obj/x86/Debug/ProntoMVC.Data.dll
	Data/obj/x86/Debug/ProntoMVC.Data.pdb
	ProntoMVC/Reportes/Proyectos de reporting/Report Project1/bin/Debug/build.obj
	ProntoMVC/obj/Debug/ProntoMVC.csproj.FileListAbsolute.txt
	TestProject2/App.config
	TestProject2/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache

Former-commit-id: 991b88e042d852697e315c96f4b0cdff92a58bb3

+++ Data/Models/Factura.cs
@@ -142,7 +142,7 @@ namespace ProntoMVC.Data.Models
         public string CuitClienteExterno { get; set; }
         public Nullable<decimal> ComisionDiferenciada { get; set; }
         public Nullable<int> IdTipoNegocioVentas { get; set; }
-        public Nullable<int> NumeroOrdenCompraExterna { get; set; }
+        public Nullable<long> NumeroOrdenCompraExterna { get; set; }
         public Nullable<int> IdTipoOperacion { get; set; }
         public string FueEnviadoCorreoConFacturaElectronica { get; set; }
         public Nullable<int> IdClienteObservaciones { get; set; }
